requirements/dev.txt
-r common.txt
django-debug-toolbar==0.10.2